pyproject.toml
[tool.poetry]
name = "telegram_bot"
version = "0.1.0"
description = ""
authors = ["Nikita Tolkachev <tolkachev.nick@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = ">=2.4.2,<2.5.0"
pydantic-settings = ">=2.0.3,<2.1.0"
asyncpg = ">=0.28,<1.0"
pytz = "2023.3.post1"
faker = "^19.13.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
freezegun = "^1.2.2"
pytest-freezegun = "^0.4.2"
python-telegram-bot = "^20.6"
aiohttp = ">=3.8,<4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
